+++ R/weights-kernel.R
@@ -32,8 +32,10 @@ st_kernel_weights <- function(nb, geometry, kernel = "uniform",
   # if self isnt included elicit warning
   self_included <- !is.null(attr(nb, "self.included"))
   if (!self_included) {
-    cli::cli_warn("It is recommended to include the ith observation.
-                    Consider `nb = include_self(nb)`.")
+    cli::cli_alert_info(
+      c("It is recommended to include the ith observation.",
+                    " Consider `nb = include_self(nb)`.")
+      )
   }
 
   # verify kernels